Prep for multiple code locations
