+++ src/mozilla_sec_eia/library/model_jobs.py
@@ -1,0 +1,88 @@
+"""Implement helper methods for constructing dagster jobs.
+
+Methods defined here are the main interface for constructing PUDL model jobs.
+`create_production_model_job` will produce a dagster job that will use the default
+multi-process executor to run a PUDL model. `create_validation_model_job` is meant for
+testing/validating models with an mlflow run backing the dagster run for logging.
+To avoid problems with mlflow runs, test/validation jobs are run with the dagster
+in process executor.
+"""
+
+import mlflow
+from dagster import (
+    AssetsDefinition,
+    HookContext,
+    JobDefinition,
+    define_asset_job,
+    failure_hook,
+    in_process_executor,
+    success_hook,
+)
+from mlflow.entities import RunStatus
+
+
+def create_production_model_job(
+    job_name: str,
+    assets: list[AssetsDefinition],
+    **kwargs,
+) -> JobDefinition:
+    """Construct a dagster job and supply Definitions with assets and resources."""
+    return define_asset_job(
+        job_name,
+        selection=assets,
+        config={
+            "ops": {},
+            "resources": {
+                "mlflow_interface": {
+                    "config": {
+                        "experiment_name": job_name,
+                        "tracking_enabled": False,
+                    }
+                }
+            },
+        },
+        **kwargs,
+    )
+
+
+@success_hook(required_resource_keys={"mlflow_interface"})
+def log_op_config(context: HookContext):
+    """Log any config supplied to ops/assets in validation job to mlflow tracking server."""
+    if context.op_config is not None:
+        mlflow.log_params(context.op_config)
+
+
+@failure_hook(required_resource_keys={"mlflow_interface"})
+def end_run_on_failure(context: HookContext):
+    """Inform mlflow about job failure."""
+    if isinstance(context.op_exception, KeyboardInterrupt):
+        mlflow.end_run(status=RunStatus.to_string(RunStatus.KILLED))
+    else:
+        mlflow.end_run(status=RunStatus.to_string(RunStatus.FAILED))
+
+
+def create_validation_model_job(
+    job_name: str,
+    assets: list[AssetsDefinition],
+    **kwargs,
+):
+    """Construct a dagster job and supply Definitions with assets and resources."""
+    return define_asset_job(
+        job_name,
+        selection=assets,
+        executor_def=in_process_executor,
+        hooks={log_op_config, end_run_on_failure},
+        # Configure mlflow_interface for job with appropriate experiment name
+        config={
+            "ops": {},
+            "resources": {
+                "mlflow_interface": {
+                    "config": {
+                        "experiment_name": job_name,
+                        "tracking_enabled": True,
+                    }
+                }
+            },
+        },
+        **kwargs,
+    )
